+++ tests/testthat/test_SignaturePresenceTest.R
@@ -10,16 +10,19 @@ test_that("Use signature id in SignaturePresenceTest", {
   sigs.to.test <- sigs[, names(sigs.prop), drop = FALSE]
   test.out <- SignaturePresenceTest(spectra = spectra,
                                     sigs = sigs.to.test,
-                                    target.sig.index = "SBS22")
+                                    target.sig.index = "SBS22",
+                                    seed = 2892)
   
   test.out1 <- SignaturePresenceTest(spectra = spectra,
                                      sigs = sigs.to.test,
-                                     target.sig.index = 13)
+                                     target.sig.index = 13,
+                                     seed = 2892)
   
   test.out2 <- 
-    expect_error(SignaturePresenceTest(spectra = spectra,
+    expect_warning(SignaturePresenceTest(spectra = spectra,
                                        sigs = sigs.to.test,
-                                       target.sig.index = "SBS10a"))
-  expect_equal(test.out[[1]]$statistic, 156.1548, tolerance = 1e-5)
+                                       target.sig.index = "SBS10a",
+                                       seed = 2892))
+  expect_equal(test.out[[1]]$statistic, 155.937, tolerance = 1e-5)
   expect_equal(test.out, test.out1)
 })